refactor(web): extract shared last-run UI into LastRun.tsx (#690 m1)
Move LastRunOutcome, LastFireDetail and their helpers (OutcomeBadge, Tally,
IssueRef, SkipRow), the SKIP_REASON_TONES map, and formatStamp out of
Schedules.tsx into a new components/LastRun.tsx, so DefaultJobs and
ScheduleGroupRow can reuse them without the circular import that re-exporting
from Schedules.tsx would create. Behavior-preserving; dedupes DefaultJobs'
local formatStamp.
